Move `impl_array_newtype_stringify` to `internals`
Another step towards #5333. Here we move the macro to internals so both
`p2p` and `bitcoin` may depend on it, with no dependency on each other.
Because `p2p` does not (...will not?) have a `serde` dependency, the
macro is decomposed into two parts. One macro implements `hex` related
traits, the other implements `serde`.
